refactor(allocator): rename `bump` module to `arena` (#21396)
Pure refactor. Rename `bump.rs` to `arena.rs`, and rename `bump` tests directory to `arena` - in line with renaming `Bump` to `Arena` (#21394).

No substantive changes, only renaming.
