+++ crates/oxc_allocator/tests/arena/capacity.rs
@@ -1,4 +1,4 @@
-use oxc_allocator::bump::Arena;
+use oxc_allocator::arena::Arena;
 
 #[test]
 fn try_with_capacity_too_large() {
