+++ crates/oxc_allocator/tests/arena/quickcheck.rs
