+++ crates/oxc_allocator/tests/arena/tests.rs
@@ -1,4 +1,4 @@
-use oxc_allocator::bump::Arena;
+use oxc_allocator::arena::Arena;
 use std::alloc::Layout;
 use std::fmt::Debug;
 use std::mem;
